test: FakeStreamStorageRegistry
